requirements_dev.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
tensorflow
tensorflow-datasets
tensorflow-probability
dataclasses
dill
Flask
Flask-Cors
dvc
notebook
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scipy
pytest==7.1.3
tox
black==22.8.0
flake8==5.0.4
mypy==0.971
pymongo
pymongo[srv]
dnspython
kaggle
missingno
mlflow==2.5.0
-e .